+++ app/src/main/AndroidManifest.xml
@@ -93,37 +93,37 @@
             android:launchMode="singleTop"
             android:screenOrientation="portrait" />
 
-        <activity android:name=".activities.brapi.update.BrapiStudyFilterActivity"
+        <activity android:name=".activities.brapi.io.BrapiStudyFilterActivity"
             android:configChanges="orientation|keyboardHidden|screenSize|locale"
             android:launchMode="singleTop"
             android:screenOrientation="portrait" />
 
-        <activity android:name=".activities.brapi.update.ListFilterActivity"
+        <activity android:name=".activities.brapi.io.ListFilterActivity"
             android:configChanges="orientation|keyboardHidden|screenSize|locale"
             android:launchMode="singleTop"
             android:screenOrientation="portrait" />
 
-        <activity android:name=".activities.brapi.update.BrapiProgramFilterActivity"
+        <activity android:name=".activities.brapi.io.BrapiProgramFilterActivity"
             android:configChanges="orientation|keyboardHidden|screenSize|locale"
             android:launchMode="singleTop"
             android:screenOrientation="portrait" />
 
-        <activity android:name=".activities.brapi.update.BrapiSeasonsFilterActivity"
+        <activity android:name=".activities.brapi.io.BrapiSeasonsFilterActivity"
             android:configChanges="orientation|keyboardHidden|screenSize|locale"
             android:launchMode="singleTop"
             android:screenOrientation="portrait" />
 
-        <activity android:name=".activities.brapi.update.BrapiTrialsFilterActivity"
+        <activity android:name=".activities.brapi.io.BrapiTrialsFilterActivity"
             android:configChanges="orientation|keyboardHidden|screenSize|locale"
             android:launchMode="singleTop"
             android:screenOrientation="portrait" />
 
-        <activity android:name=".activities.brapi.update.BrapiCropsFilterActivity"
+        <activity android:name=".activities.brapi.io.BrapiCropsFilterActivity"
             android:configChanges="orientation|keyboardHidden|screenSize|locale"
             android:launchMode="singleTop"
             android:screenOrientation="portrait" />
 
-        <activity android:name=".activities.brapi.update.BrapiStudyImportActivity"
+        <activity android:name=".activities.brapi.io.BrapiStudyImportActivity"
             android:configChanges="orientation|keyboardHidden|screenSize|locale"
             android:launchMode="singleTop"
             android:screenOrientation="portrait" />
