+++ app/src/main/java/com/fieldbook/tracker/activities/brapi/io/TrialStudyModel.kt
@@ -1,4 +1,4 @@
-package com.fieldbook.tracker.activities.brapi.update
+package com.fieldbook.tracker.activities.brapi.io
 
 import org.brapi.v2.model.core.BrAPIStudy
