+++ app/src/main/java/com/fieldbook/tracker/dialogs/BrapiFilterBottomSheetDialog.kt
@@ -12,12 +12,12 @@ import android.widget.TextView
 import androidx.annotation.OptIn
 import androidx.appcompat.content.res.AppCompatResources
 import com.fieldbook.tracker.R
-import com.fieldbook.tracker.activities.brapi.update.BrapiCropsFilterActivity
-import com.fieldbook.tracker.activities.brapi.update.BrapiFilterTypeAdapter
-import com.fieldbook.tracker.activities.brapi.update.BrapiProgramFilterActivity
-import com.fieldbook.tracker.activities.brapi.update.BrapiSeasonsFilterActivity
-import com.fieldbook.tracker.activities.brapi.update.BrapiStudyFilterActivity
-import com.fieldbook.tracker.activities.brapi.update.BrapiTrialsFilterActivity
+import com.fieldbook.tracker.activities.brapi.io.BrapiCropsFilterActivity
+import com.fieldbook.tracker.activities.brapi.io.BrapiFilterTypeAdapter
+import com.fieldbook.tracker.activities.brapi.io.BrapiProgramFilterActivity
+import com.fieldbook.tracker.activities.brapi.io.BrapiSeasonsFilterActivity
+import com.fieldbook.tracker.activities.brapi.io.BrapiStudyFilterActivity
+import com.fieldbook.tracker.activities.brapi.io.BrapiTrialsFilterActivity
 import com.fieldbook.tracker.preferences.GeneralKeys
 import com.google.android.material.badge.ExperimentalBadgeUtils
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
