+++ app/src/main/res/layout/activity_study_importer.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent">
 
     <com.google.android.material.appbar.MaterialToolbar
-        android:id="@+id/act_brapi_importer_tb"
+        android:id="@+id/act_list_filter_tb"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="?attr/fb_color_primary"
@@ -16,12 +16,12 @@
 
     <ProgressBar style="@style/Widget.AppCompat.ProgressBar.Horizontal"
         android:visibility="visible"
-        android:id="@+id/brapi_importer_pb"
+        android:id="@+id/act_list_filter_pb"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:indeterminate="true"
         android:layout_marginTop="-7dp"
-        app:layout_constraintTop_toBottomOf="@id/act_brapi_importer_tb"
+        app:layout_constraintTop_toBottomOf="@id/act_list_filter_tb"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"/>
 
@@ -33,7 +33,7 @@
         android:gravity="center"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/brapi_importer_pb"/>
+        app:layout_constraintTop_toBottomOf="@id/act_list_filter_pb"/>
 
     <com.google.android.material.tabs.TabLayout
         android:visibility="gone"
@@ -78,14 +78,14 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@id/brapi_importer_tl"
-        app:layout_constraintBottom_toTopOf="@id/brapi_importer_rv"
+        app:layout_constraintBottom_toTopOf="@id/act_list_filter_rv"
         android:choiceMode="singleChoice"
         app:layout_constraintVertical_weight="0.25"/>
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:id="@+id/brapi_importer_rv"
+        android:id="@+id/act_list_filter_rv"
         android:orientation="vertical"
         tools:listitem="@layout/list_item_study"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
@@ -134,7 +134,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         android:text="@string/dialog_save"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/brapi_importer_rv"
+        app:layout_constraintTop_toBottomOf="@id/act_list_filter_rv"
         app:layout_constraintEnd_toEndOf="parent"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
